refactor: split monolithic test file, fix exports and CI
- Split python/test/test.py (1,563 lines) into 6 domain-specific files:
  test_time, test_coordinates, test_frames, test_ephemeris,
  test_propagation, test_sgp4
- Add conftest.py with shared testvec_dir fixture
- Update CI pipelines (build.yml, release.yml) to run pytest python/test/
- Fix __init__.pyi __all__: add missing frame, weekday, sgp4_error,
  sgp4_gravconst, sgp4_opsmode, geodetic, gravity_and_partials, nrlmsise00
- Fix misleading "Currently not used" comment on frames module in lib.rs
