+++ python/satkit/__init__.pyi
@@ -14,14 +14,22 @@ __all__ = [
     "time",
     "duration",
     "timescale",
+    "weekday",
     "quaternion",
+    "frame",
     "sgp4",
+    "sgp4_error",
+    "sgp4_gravconst",
+    "sgp4_opsmode",
     "gravmodel",
     "gravity",
+    "gravity_and_partials",
+    "nrlmsise00",
     "density",
     "solarsystem",
     "TLE",
     "itrfcoord",
+    "geodetic",
     "kepler",
     "consts",
     "frametransform",
