+++ python/test/conftest.py
@@ -1,0 +1,11 @@
+import os
+
+import pytest
+
+
+@pytest.fixture
+def testvec_dir():
+    return os.getenv(
+        "SATKIT_TESTVEC_ROOT",
+        default="." + os.path.sep + "satkit-testvecs" + os.path.sep,
+    )
